Add files via upload
Update contains :
hashtag word cloud visualisation 
popular words and hashtag word frequency rank tables
Mock up of image tree with different weights for each image dependant on "popularity" this will be linked to a modal view with the relevant comments when the comment urls from the live feed become available. I tried a multitude of different methods to stack the images in a remotely tessellating fashion from pure CSS to masonry JS but none of them gave the desired effect, so I eventually went with a similar algorithm to the word cloud to position the images in a random (but replicable position) that is aesthetically pleasing enough but also easily clickable. I hope this will suffice for this iteration.